+++ app/chat/page.tsx
@@ -3,7 +3,7 @@ import { SidebarInset } from '@/components/ui/sidebar';
 import ChatHeader from '@/components/chat/ChatHeader';
 import ChatInputBox from '@/components/chat/ChatInputBox';
 import ChatArea from '@/components/chat/ChatArea';
-import { ChatContextProvider } from '@/components/chat/ChatContextProvider';
+import ChatContextProvider from '@/components/chat/ChatContextProvider';
 import ChatSidePanel from '@/components/chat/ChatSidePanel';
 
 export default function Chat() {
