+++ src/components/graph/PhotoFolder.tsx
