+++ src/components/graph/PhotoNode.tsx
