+++ src/types/apis/apiChat.d.ts
@@ -4,7 +4,21 @@ type APIResponseMessage = {
   agent_id: string;
   agent_name: string;
   agent_role: string;
-  product_image_url: string[];
+  products: {
+    product_url: string;
+    product_id: string;
+  }[];
+};
+
+type APIRoomIdMessage = {
+  id: number;
+  content: string;
+  image_url: null;
+  message_type: string;
+  created_at: string;
+  agent_type: string | null;
+  agent_name: string | null;
+  product_image_url: string[] | null;
 };
 
 type APIRoomHistory = {
@@ -16,5 +30,12 @@ type APIRoomHistory = {
 };
 
 type APIRoomIdMessages = {
-  messages: APIMessage[];
+  messages: APIRoomIdMessage[];
+};
+
+type APIProduct = {
+  product_name: string;
+  comprehensive_description: string;
+  style_tags: string[];
+  tpo_tags: string[];
 };
